requirements.txt
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
falcon==2.0.0
Flask==1.1.2
hug==2.6.1
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
numpy==1.19.4
packaging==20.4
pdocs==1.0.2
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
python-dotenv==0.15.0
requests==2.25.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
